+++ src/Leetcode_Arrays/DecompressRunLengthEncodedList.java
@@ -1,4 +1,4 @@
-package Leetcode;
+package Leetcode_Arrays;
 
 import java.util.LinkedList;
 import java.util.List;
