+++ src/Leetcode_Arrays/KidsWithGreatestNumberCandies.java
@@ -1,4 +1,4 @@
-package Leetcode;
+package Leetcode_Arrays;
 
 import java.util.Arrays;
 import java.util.List;
