+++ src/Leetcode_Arrays/MaximumNumberOfWordsFoundInSentences.java
@@ -1,4 +1,4 @@
-package Leetcode;
+package Leetcode_Arrays;
 
 public class MaximumNumberOfWordsFoundInSentences {
