+++ src/Leetcode_Arrays/MinimumNumberMoves.java
@@ -1,4 +1,4 @@
-package Leetcode;
+package Leetcode_Arrays;
 
 import java.util.Arrays;
